requirements.txt
Adafruit-Blinka==6.10.1
adafruit-circuitpython-neopixel==6.1.3
adafruit-circuitpython-pypixelbuf==2.2.5
Adafruit-PlatformDetect==3.13.4
Adafruit-PureIO==1.1.8
bidict==0.21.2
certifi==2021.5.30
chardet==4.0.0
idna==2.10
pkg-resources==0.0.0
pyftdi==0.53.1
pyserial==3.5
python-engineio==4.2.0
python-socketio==5.3.0
pyusb==1.1.1
requests==2.25.1
rpi-ws281x==4.3.0
RPi.GPIO==0.7.0
sysv-ipc==1.1.0
urllib3==1.26.5
websocket-client==1.0.1
